release zarr version constraint (#283)
Allow zarr v3 to be used and should be compatible with the toolbox
