+++ pyproject.toml
@@ -16,7 +16,7 @@ requests = ">=2.27.1"
 setuptools = ">=68.2.2"
 xarray = ">=2023.4.0"
 tqdm = ">=4.65.0"
-zarr = "^2.13.3"
+zarr = ">=2.13.3"
 dask = ">=2022"
 boto3 = ">=1.26"
 semver = ">=3.0.2"
@@ -25,7 +25,7 @@ lxml = ">=4.9.0"
 numpy = ">=1.23.0"
 pydantic = "^2.9.1"
 h5netcdf = "^1.4.0"
-arcosparse = "^0.3.1"
+arcosparse = "^0.3.2"
 
 [project.scripts]
 copernicusmarine = 'copernicusmarine.command_line_interface.copernicus_marine:command_line_interface'
